requirements.txt
torch
numpy
pyedflib
pandas
# local package
-e .

# external requirements
click
Sphinx
coverage
flake8
python-dotenv>=0.5.1
